Moved tasks to Tasks.c
